requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
pillow==10.2.0
pytz==2024.1
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
django-probes==1.7.0
python-dotenv==1.0.1
setuptools==69.2.0

django-smart-selects~=1.6.0